refactor: 리팩토링 본문 참조
- constans 패키지 안에 event 패키지 생성 후 클래스 이동
- Gift enum 내부 count 필드 생성
