+++ src/test/java/christmas/domain/EventTest.java
@@ -1,6 +1,6 @@
 package christmas.domain;
 
-import christmas.domain.contants.EventDiscount;
+import christmas.domain.contants.event.EventDiscount;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
